% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densenet.R
\name{conv_factory}
\alias{conv_factory}
\title{Apply BtachNorm, Relu, 3x3Conv2D, optional dropout}
\usage{
conv_factory(x, nb_filter, dropout_rate = NULL, weight_decay = 1e-04)
}
\arguments{
\item{x}{Input keras network}

\item{nb_filter}{int -- number of filters}

\item{dropout_rate}{double -- dropout rate}

\item{weight_decay}{double -- weight decay factor}
}
\value{
keras network with b_norm, relu and convolution2d added
}
\description{
Apply BtachNorm, Relu, 3x3Conv2D, optional dropout
}
