% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densenet.R
\name{denseblock}
\alias{denseblock}
\title{Build a denseblock where the output of each conv_factory is fed to
subsequent ones}
\usage{
denseblock(x, nb_layers, nb_filter, growth_rate)
}
\description{
Build a denseblock where the output of each conv_factory is fed to
subsequent ones
}
