% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densenet.R
\name{densenet}
\alias{densenet}
\title{Build the DenseNet model}
\usage{
densenet(nb_classes, img_dim, depth, nb_dense_block, growth_rate, nb_filter,
  dropout_rate = NULL, weight_decay = 1e-04)
}
\arguments{
\item{nb_classes}{numeric -- number of classes}

\item{img_dim}{list -- (channels, rows, colums)}

\item{depth}{numeric -- how many layers}

\item{nb_dense_block}{numeric -- number of dense blocks to add to end}

\item{growth_rate}{numeric -- number of filters to add}

\item{nb_filter}{numeric -- number of filters}

\item{dropout_rate}{numeric -- dropout rate}

\item{weight_decay}{numeric -- weight decay}
}
\description{
Build the DenseNet model
}
