% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densenet.R
\name{transition}
\alias{transition}
\title{Apply BatchNorm, Relu, 1x1Conv2D, optional dropout and Maxpooling2D}
\usage{
transition(x, nb_filter, dropout_rate = NULL, weight_decay = 1e-04)
}
\arguments{
\item{x}{Input keras network}

\item{nb_filter}{int -- number of filters}

\item{dropout_rate}{double -- dropout rate}

\item{weight_decay}{double -- weight decay factor}
}
\value{
keras network with b_norm, relu, convolution2d and maxpooling added
}
\description{
Apply BatchNorm, Relu, 1x1Conv2D, optional dropout and Maxpooling2D
}
