% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densenet.R
\name{transition_up_block}
\alias{transition_up_block}
\title{SubpixelConvolutional Upscaling (factor = 2)}
\usage{
transition_up_block(ip, nb_filters, type = "upsampling",
  weight_decay = 1e-04)
}
\arguments{
\item{ip}{keras tensor}

\item{nb_filters}{number of layers}

\item{type}{can be 'upsampling', 'subpixel', 'deconv'. Determines type of
upsampling performed}

\item{weight_decay}{weight decay factor}
}
\description{
SubpixelConvolutional Upscaling (factor = 2)
}
\seealso{
Other internal: \code{\link{conv_block}},
  \code{\link{create_dense_net}}, \code{\link{dense_block}}
}
