% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densenet2.R
\name{conv_block}
\alias{conv_block}
\title{Apply BatchNorm, Relu, 3x3 Conv2D, optional bottleneck block and dropout}
\usage{
conv_block(ip, nb_filter, bottleneck = FALSE, dropout_rate = NULL,
  weight_decay = 1e-04)
}
\arguments{
\item{ip:}{Input keras tensor}

\item{nb_filter:}{number of filters}

\item{bottleneck:}{add bottleneck block}

\item{dropout_rate:}{dropout rate}

\item{weight_decay:}{weight decay factor}
}
\description{
Apply BatchNorm, Relu, 3x3 Conv2D, optional bottleneck block and dropout
}
\references{
https://github.com/titu1994/DenseNet/blob/master/densenet.py
}
